On branch main

+++ src/api/notes/handler.js
@@ -1,38 +1,37 @@
-const ClientError = require("../../exceptions/ClientError");
+const autoBind = require('auto-bind');
+const ClientError = require('../../exceptions/ClientError');
 
 
 class NotesHandler {
   constructor(service, validator){
-    this._service = service;
+    this._service = service
     this._validator = validator;
 
-    this.postNoteHandler = this.postNoteHandler.bind(this);
-    this.getNotesHandler = this.getNotesHandler.bind(this);
-    this.getNoteByIdHandler = this.getNoteByIdHandler.bind(this);
-    this.putNoteByIdHandler = this.putNoteByIdHandler.bind(this);
-    this.deleteNoteByIdHandler = this.deleteNoteByIdHandler.bind(this);
-
+    autoBind(this)
   }
 
   async postNoteHandler(request, h){
-      this._validator.validateNotePayload(request.payload);
-      const { title = 'untitled', tags, body } = request.payload;
+    this._validator.validateNotePayload(request.payload);
+    const { title = 'untitled', tags, body } = request.payload;
+    const credentialId  = request.auth.credentials.id
 
-      const noteId =await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({title, body, tags}, credentialId);
 
-      const response = h.response({
-        status: 'success',
-        message: 'Catatan berhasil ditambahkan',
-        data: {
-          noteId, 
-        },
-      });
-      response.code(201);
-      return response;
+    const response = h.response({
+      status: 'success',
+      message: 'Catatan berhasil ditambahkan',
+      data: {
+        noteId,
+      },
+    });
+    response.code(201);
+    return response;
   }
 
   async getNotesHandler(request, h){
-    const notes = await this._service.getNote();
+    const {id : credentialsId} = request.auth.credentials
+
+    const notes = await this._service.getNote(credentialsId);
     const response = h.response({
       status : 'success',
       data : {
@@ -44,39 +43,49 @@ class NotesHandler {
   }
 
   async getNoteByIdHandler(request, h){
-      const { id } = request.params;
-      const note = await this._service.getNoteById(id);
-      const response = h.response({
-        status : 'success',
-        data : {
-          note
-        }
-      }).code(200);
-      return response;
+    const { id } = request.params;
+    const { id : credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId)
+    const note = await this._service.getNoteById(id);
+
+
+    const response = h.response({
+      status : 'success',
+      data : {
+        note
+      }
+    }).code(200);
+    return response;
   }
 
   async putNoteByIdHandler(request, h){
-      this._validator.validateNotePayload(request.payload); 
-      const { title, tags, body } = request.payload;
-      const { id } = request.params;
+    this._validator.validateNotePayload(request.payload);
+    const { title, tags, body } = request.payload;
+    const { id } = request.params;
+    const {id : credentialsId} = request.auth.credentials
 
-      await this._service.editNoteById(id, { title, tags, body });
+    await this._service.verifyNoteOwner(id, credentialsId)
+    await this._service.editNoteById(id, { title, tags, body });
 
-      const response = h.response({
-        status : 'success',
-        message : 'Catatan berhasil diperbarui'
-      }).code(200);
-      return response;
+    const response = h.response({
+      status : 'success',
+      message : 'Catatan berhasil diperbarui'
+    }).code(200);
+    return response;
   }
 
   async deleteNoteByIdHandler(request, h){
     const { id } = request.params;
-      await this._service.deleteNoteById(id);
-      const response = h.response({
-        status : 'success',
-        message : 'Catatan berhasil dihapus'
-      }).code(200);
-      return response;
+    const {id : credentialsId} = request.auth.credentials
+    
+    await this._service.verifyNoteOwner(id, credentialsId)
+    await this._service.deleteNoteById(id);
+    const response = h.response({
+      status : 'success',
+      message : 'Catatan berhasil dihapus'
+    }).code(200);
+    return response;
   }
 }
